move file splash and welocome to out
